app/requirements.txt
flask
gunicorn
pandas
numpy
google-cloud-storage==1.6.0
lightfm
